Move tests to packages matching production
