+++ generator/src/test/java/com/scottlogic/deg/generator/constraints/atomic/IsInSetConstraintTests.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.constraints;
+package com.scottlogic.deg.generator.constraints.atomic;
 
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.constraints.atomic.IsInSetConstraint;
